requirements.txt
requests >= 2.25.1
PyNaCl >= 1.4
paramiko >= 2.7.2
